Listener

+++ 后端/访问数量记录及表格展示/chart/src/Entity/IP.java
@@ -22,7 +22,7 @@ public class IP {
         this.ip = ip;
     }
 
-    public long getNumber() {
+    public int getNumber() {
         return number;
     }
